fj_core

Module algorithms

Collection of algorithms that operate on geometry

Algorithmic code is collected in this module, to keep other modules focused on their respective purpose.

Implementation Note

This module exists in a bit of an in-between state, as some of the things that are still here are probably better placed in the operations module.

Modules

approx

Approximation of objects

bounding_volume

Compute a bounding volume for an object

intersect

Intersection algorithms

triangulate

Shape triangulation