fj_core::algorithms::approx

Trait Approx

pub trait Approx: Sized {
    type Approximation;
    type Cache: Default;

    // Required method
    fn approx_with_cache(
        self,
        tolerance: impl Into<Tolerance>,
        cache: &mut Self::Cache,
        geometry: &Geometry,
    ) -> Self::Approximation;

    // Provided method
    fn approx(
        self,
        tolerance: impl Into<Tolerance>,
        geometry: &Geometry,
    ) -> Self::Approximation { ... }
}

Approximate an object

Required Associated Types

type Approximation

The approximation of the object

type Cache: Default

The cache used to cache approximation results

Required Methods

fn approx_with_cache( self, tolerance: impl Into<Tolerance>, cache: &mut Self::Cache, geometry: &Geometry, ) -> Self::Approximation

Approximate the object, using the provided cache

This is a lower-level method that allows some degree of control over caching. Callers might consider using Approx::approx instead.

Provided Methods

fn approx( self, tolerance: impl Into<Tolerance>, geometry: &Geometry, ) -> Self::Approximation

Approximate the object

tolerance defines how far the approximation is allowed to deviate from the actual object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.