fj_core::geometry

Struct Geometry

pub struct Geometry { /* private fields */ }

Geometric data that is associated with topological objects

Implementations

impl Geometry

pub fn new(topology: &Topology) -> Self

Create a new instance of Geometry

pub fn of_curve(&self, curve: &Handle<Curve>) -> Option<&CurveGeom>

Access the geometry of the provided curve

pub fn of_surface(&self, surface: &Handle<Surface>) -> &SweptCurve

Access the geometry of the provided surface

Panics

Panics, if the geometry of the surface is not defined.

pub fn of_vertex(&self, vertex: &Handle<Vertex>) -> Option<&VertexGeom>

Access the geometry of the provided vertex

pub fn of_curve_2(&self, curve: &Handle<Curve>) -> Option<&CurveGeom2>

Access the geometry of the provided curve

Implementation Note

There currently is an ongoing transition to a new geometry system. This method returns new-style geometry. Its name is temporary, while the method returning the old-style geometry is still taking up the more concise name.

pub fn of_surface_2(&self, surface: &Handle<Surface>) -> Option<&SurfaceGeom>

Access the geometry of the provided surface

Implementation Note

There currently is an ongoing transition to a new geometry system. This method returns new-style geometry. Its name is temporary, while the method returning the old-style geometry is still taking up the more concise name.

pub fn xy_plane(&self) -> &SweptCurve

Access the geometry of the xy-plane

pub fn xz_plane(&self) -> &SweptCurve

Access the geometry of the xz-plane

pub fn yz_plane(&self) -> &SweptCurve

Access the geometry of the yz-plane