fj_core::geometry::traits

Trait GenPolyline

pub trait GenPolyline<const D: usize> {
    // Required methods
    fn origin(&self) -> Point<D>;
    fn line_segment_at(
        &self,
        point_curve: Point<1>,
        tolerance: Tolerance,
    ) -> LineSegment<D>;
    fn generate_polyline(
        &self,
        boundary: CurveBoundary<Point<1>>,
        tolerance: Tolerance,
    ) -> Vec<Point<1>>;
}

Generate polylines, the uniform representation of curve geometry

This trait provides a generic and uniform interface to curve geometry. It is implemented by types that represent specific kinds of curve geometry.

It is generic over the dimensionality of the generated polyline. Typically, two variants should be implemented per curve geometry type:

GenPolyline<2> for surface-local geometry.

GenPolyline<3> for global 3D geometry.

Required Methods

fn origin(&self) -> Point<D>

Access the origin of the curve

fn line_segment_at( &self, point_curve: Point<1>, tolerance: Tolerance, ) -> LineSegment<D>

Compute a line segment to approximate the curve at this point

Degenerate Case

If the curve requires no approximation (meaning it is a line), then per convention, a degenerate line segment is returned, that collapses to the provided point.

fn generate_polyline( &self, boundary: CurveBoundary<Point<1>>, tolerance: Tolerance, ) -> Vec<Point<1>>

Generate a polyline within the provided boundary