fj_core::geometry::traits

Trait GenTriMesh

pub trait GenTriMesh {
    // Required methods
    fn origin(&self, geometry: &Geometry) -> Point<3>;
    fn triangle_at(
        &self,
        point_surface: Point<2>,
        tolerance: Tolerance,
        geometry: &Geometry,
    ) -> (Triangle<3>, [Scalar; 3]);
    fn generate_tri_mesh(
        &self,
        boundary: Aabb<2>,
        tolerance: Tolerance,
        geometry: &Geometry,
    ) -> Vec<Point<2>>;
}

Generate triangle meshes, the uniform representation of surface geometry

Required Methods

fn origin(&self, geometry: &Geometry) -> Point<3>

Access the origin of the surface

fn triangle_at( &self, point_surface: Point<2>, tolerance: Tolerance, geometry: &Geometry, ) -> (Triangle<3>, [Scalar; 3])

Return the triangle at the provided point on the surface

Select a triangle of the surface’s triangle mesh representation, the one at the provided surface point. Return that triangle, as well as the barycentric coordinates of the provided point on the triangle.

Triangle Size and Validity

If a surface is curved along both axes, the triangle’s size is chosen such, that it approximates the surface, with the maximum allowed deviation of the actual surface defined by the provided tolerance argument.

Otherwise, the size of the returned triangle is at least partially arbitrary. Take the extreme case of a plane: Since it is not curved at all, the returned triangle can be arbitrarily large.

However, since surfaces are infinite, and we can’t represent infinite triangles, there is no sensible upper bound for the size. Instead, to prevent an arbitrary choice for the size of triangles, which would imply properties of the surface that are not true, and might therefore be confusing, the triangles returned by this function have a length of zero along axes that do not require approximation.

The most extreme case would be a plane, for which the returned triangle is collapsed to a point. For a cylinder, the triangle would have the appropriate width to approximate the curved axis given the provided tolerance, while having zero height.

Implementation Note

At the time this was written, there was no dedicated type to represent barycentric coordinates. Nor any other code that used them, I think.

If this changes, and a special type for barycentric coordinates is added, it would make sense to return that here.

fn generate_tri_mesh( &self, boundary: Aabb<2>, tolerance: Tolerance, geometry: &Geometry, ) -> Vec<Point<2>>

Generated a triangle mesh within the provided boundary