fj_core::layers

Struct Layer

pub struct Layer<S> { /* private fields */ }

A generic layer, which controls access to layer state

Layer is a generic wrapper around some state and controls access to it. It Derefs to the state it wraps, for easy read access, but prevents any direct write access.

Instead, each write access to state is reified as a command, which are processed by Layer::process_command. Processing a command can result in any number of events, which can then be used as commands for other layers.

This design takes inspiration from, and uses the nomenclature of, this article: https://thinkbeforecoding.com/post/2021/12/17/functional-event-sourcing-decider

Implementations

impl Layer<Geometry>

pub fn define_curve( &mut self, curve: Handle<Curve>, surface: Handle<Surface>, geometry: LocalCurveGeom, )

Define the geometry of the provided curve

pub fn define_curve_2(&mut self, curve: Handle<Curve>, geometry: CurveGeom2)

Define the geometry of the provided curve

Implementation Note

There currently is an ongoing transition to a new geometry system. This method defines new-style geometry. Its name is temporary, while the method defining the old-style geometry is still taking up the more concise name.

pub fn define_surface(&mut self, surface: Handle<Surface>, geometry: SweptCurve)

Define the geometry of the provided surface

Panics

Panics, if the surface is a special pre-defined plane, like the basis planes (xy-, xz-, or yz-plane).

pub fn define_surface_2( &mut self, surface: Handle<Surface>, geometry: SurfaceGeom, )

Define the geometry of the provided surface

Panics

Panics, if the surface is a special pre-defined plane, like the basis planes (xy-, xz-, or yz-plane).

Implementation Note

There currently is an ongoing transition to a new geometry system. This method defines new-style geometry. Its name is temporary, while the method defining the old-style geometry is still taking up the more concise name.

pub fn define_vertex( &mut self, vertex: Handle<Vertex>, curve: Handle<Curve>, geometry: LocalVertexGeom, )

Define the geometry of the provided vertex

impl Layer<Presentation>

pub fn set_color(&mut self, region: Handle<Region>, color: Color)

Set the color of a region

pub fn derive_object( &mut self, original: AnyObject<Stored>, derived: AnyObject<Stored>, )

Mark an object as being derived from another

impl Layer<Topology>

pub fn insert( &mut self, object: AnyObject<AboutToBeStored>, geometry: &mut Layer<Geometry>, validation: &mut Layer<Validation>, )

Insert an object into the stores

Passes any events produced to the validation layer.

impl Layer<Validation>

pub fn take_errors(&mut self) -> Result<(), ValidationErrors>

Take all errors stored in the validation layer

impl<S> Layer<S>

pub fn new(state: S) -> Self

Create an instance of Layer

pub fn process_command<C>(&mut self, command: C) -> C::Result
where C: Command<S>,

Process a command without capturing any events

The command is processed synchronously. When this method returns, the state has been updated.

pub fn process_command_and_capture_events<C>( &mut self, command: C, events: &mut Vec<C::Event>, ) -> C::Result
where C: Command<S>, C::Event: Clone,

Process a command and capture the events that produces

The command is processed synchronously. When this method returns, the state has been updated.

pub fn into_state(self) -> S

Drop this instance, returning the wrapped state