fj_core::operations

Module build

Operations to build objects

All object structs have constructors, but building even simple shapes by just using those is pretty difficult. This is why the traits in this module exist. They build on top of those low-level constructors, providing a more rich set of operations to build geometry.

Wrapper Structs

Many of the the trait methods return the object that is being built directly, but others return a wrapper struct. An example of this is BuildFace::triangle, which returns Polygon instead of returning the face directly.

These wrapper structs are designed to provide convenient access not only to the top-level object itself, but also to the other objects that make up its components.

Structs

Polygon

A polygon

Tetrahedron

A tetrahedron

TetrahedronShell

A tetrahedron

Traits

BuildCurve

Build a Curve

BuildCycle

Build a Cycle

BuildFace

Build a Face

BuildHalfEdge

Build a HalfEdge

BuildRegion

Build a Region

BuildShell

Build a Shell

BuildSketch

Build a Sketch

BuildSolid

Build a Solid

BuildSurface

Build a Surface