fj_core::operations::build

Trait BuildCycle

pub trait BuildCycle {
    // Provided methods
    fn empty() -> Cycle { ... }
    fn from_half_edges_and_boundaries<I>(
        half_edges_and_boundaries: I,
        core: &mut Core,
    ) -> Cycle
       where I: IntoIterator<Item = (Handle<HalfEdge>, CurveBoundary<Point<1>>)>,
             I::IntoIter: Clone + ExactSizeIterator { ... }
    fn circle(
        center: impl Into<Point<2>>,
        radius: impl Into<Scalar>,
        surface: Handle<Surface>,
        core: &mut Core,
    ) -> Cycle { ... }
    fn polygon<P, Ps>(
        points: Ps,
        surface: Handle<Surface>,
        core: &mut Core,
    ) -> Cycle
       where P: Into<Point<2>>,
             Ps: IntoIterator<Item = P>,
             Ps::IntoIter: Clone + ExactSizeIterator { ... }
}

Build a Cycle

See module-level documentation for context.

Provided Methods

fn empty() -> Cycle

Build an empty cycle

fn from_half_edges_and_boundaries<I>( half_edges_and_boundaries: I, core: &mut Core, ) -> Cycle
where I: IntoIterator<Item = (Handle<HalfEdge>, CurveBoundary<Point<1>>)>, I::IntoIter: Clone + ExactSizeIterator,

Build a cycle from half-edges and associated curve boundaries

fn circle( center: impl Into<Point<2>>, radius: impl Into<Scalar>, surface: Handle<Surface>, core: &mut Core, ) -> Cycle

Build a circle

This circle is built out of 4 distinct arcs.

Implementation Note

The cycle can’t be built out of a single half-edge. That would be invalid although there’s no validation check to document and enforce that yet. Please refer to the following issue for more information: https://github.com/hannobraun/fornjot/issues/2374

The cycle is built out of 4 arcs specifically, because that’s easier to implement than three, and building it out of two creates geometry that the cycle winding code can’t handle right now. The following issue has more information on the cycle winding problems: https://github.com/hannobraun/fornjot/issues/2130

fn polygon<P, Ps>( points: Ps, surface: Handle<Surface>, core: &mut Core, ) -> Cycle
where P: Into<Point<2>>, Ps: IntoIterator<Item = P>, Ps::IntoIter: Clone + ExactSizeIterator,

Build a polygon

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.