fj_core::operations::insert

Trait IsInserted

pub trait IsInserted {
    type T<T>: Borrow<T>;
}

Indicate whether an object has been inserted

Intended to be used as a type parameter bound for structs that need to track whether their contents have been inserted or not.

Required Associated Types

type T<T>: Borrow<T>

The type of the object for which the insertion status is tracked

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.