fj_core::operations::join

Trait JoinCycle

pub trait JoinCycle {
    // Required methods
    fn add_joined_half_edges<Es>(
        &self,
        half_edges: Es,
        surface: Handle<Surface>,
        core: &mut Core,
    ) -> Self
       where Es: IntoIterator<Item = (Handle<HalfEdge>, LocalCurveGeom)>,
             Es::IntoIter: Clone + ExactSizeIterator;
    fn join_to(
        &self,
        other: &Cycle,
        range: RangeInclusive<usize>,
        other_range: RangeInclusive<usize>,
        surface_self: Handle<Surface>,
        core: &mut Core,
    ) -> Self;
}

Join a Cycle to another

Required Methods

fn add_joined_half_edges<Es>( &self, half_edges: Es, surface: Handle<Surface>, core: &mut Core, ) -> Self
where Es: IntoIterator<Item = (Handle<HalfEdge>, LocalCurveGeom)>, Es::IntoIter: Clone + ExactSizeIterator,

Add new half-edges to the cycle that are joined to the provided ones

This method creates a new half-edge for each half-edge that is provided, joins the new half-edge to the provided one, and adds the new half-edge to the cycle.

Expects the provided half-edges to be in an unnatural order:

This is the opposite order that they would appear in, in a cycle.

Meaning each half-edge ends where the previous one starts.

The geometry for each new half-edge needs to be provided as well. Also requires the surface that the cycle is defined in.

fn join_to( &self, other: &Cycle, range: RangeInclusive<usize>, other_range: RangeInclusive<usize>, surface_self: Handle<Surface>, core: &mut Core, ) -> Self

Join the cycle to another

Joins the cycle to the other at the provided ranges. The ranges specify the indices of the edges that are joined together.

A modulo operation is applied to all indices before use, so in a cycle of 3 edges, indices 0 and 3 refer to the same edge. This allows for specifying a range that crosses the “seam” of the cycle.

Panics

Panics, if the ranges have different lengths.

Assumptions

This method makes some assumptions that need to be met, if the operation is to result in a valid shape:

The joined edges must be coincident.

The locally defined curve coordinate systems of the edges must match.

If either of those assumptions are not met, this will result in a validation error down the line.

Implementation Note

The use of the RangeInclusive type might be a bit limiting, as other range types might be more convenient in a given use case. This implementation was chosen for now, as it wasn’t clear whether the additional complexity of using RangeBounds would be worth it.

A solution based on SliceIndex could theoretically be used, but that trait is partially unstable. In addition, it’s not clear how it could be used generically, as it could yield a range (which can be iterated over) or a single item (which can not). This is not a hard problem in principle (a single item could just be an iterator of length 1), but I don’t see it how to address this in Rust in a reasonable way.

Maybe a custom trait that is implemented for usize and all range types would be the best solution.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.