fj_core::operations::reverse

Trait ReverseCurveCoordinateSystems

pub trait ReverseCurveCoordinateSystems {
    type Reversed;

    // Required method
    fn reverse_curve_coordinate_systems(self, core: &mut Core) -> Self::Reversed;
}

Reverse the direction of the curve coordinate systems within an object

Required Associated Types

type Reversed

The type of the reversed object

Required Methods

fn reverse_curve_coordinate_systems(self, core: &mut Core) -> Self::Reversed

Reverse the direction of the curve coordinate systems within an object

This will not have any effect on object positions in global coordinates.