fj_core::operations

Module sweep

Sweep objects along a path to create new objects

Sweeps 1D or 2D objects along a straight path, creating a 2D or 3D object, respectively.

Structs

ShellExtendedBySweep

The result of sweeping a Face of a Shell

SweepCache

A cache used for sweeping

SweptCycle

The result of sweeping a Cycle

SweptHalfEdge

The result of sweeping a HalfEdge

SweptRegion

The result of sweeping a Region

Traits

SweepCycle

Sweep a Cycle

SweepFace

Sweep a Face

SweepFaceOfShell

Sweep a Face that is part of a Shell

SweepHalfEdge

Sweep a HalfEdge

SweepRegion

Sweep a Region

SweepSketch

Sweep a Sketch

SweepSurfacePath

Sweep a Path

SweepVertex

Sweep a Vertex