fj_core::operations::sweep

Trait SweepCycle

pub trait SweepCycle {
    // Required method
    fn sweep_cycle(
        &self,
        bottom_surface: Handle<Surface>,
        top_surface: Handle<Surface>,
        color: Option<Color>,
        path: impl Into<Vector<3>>,
        cache: &mut SweepCache,
        core: &mut Core,
    ) -> SweptCycle;
}

Sweep a Cycle

See module documentation for more information.

Required Methods

fn sweep_cycle( &self, bottom_surface: Handle<Surface>, top_surface: Handle<Surface>, color: Option<Color>, path: impl Into<Vector<3>>, cache: &mut SweepCache, core: &mut Core, ) -> SweptCycle

Sweep the Cycle

Sweep the cycle into a set of connected faces. Each half-edge in the cycle is swept into a face, meaning all resulting faces form a connected set of side walls.

Requires the surface that the half-edges of the cycle are defined in, and optionally the color of the created faces.

There is no face at the “top” (the end of the sweep path), as we don’t have enough information here to create that. We just have a single cycle, and don’t know whether that is supposed to be the only cycle within a potential top face, or whether it’s an exterior or interior one.

For the same reason, there also is no “bottom” face. Additionally, whether a bottom face is even desirable depends on the context this operation is called in, and therefore falls outside of its scope.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.