fj_core::operations::sweep

Trait SweepFaceOfShell

pub trait SweepFaceOfShell {
    // Required method
    fn sweep_face_of_shell(
        &self,
        face: Handle<Face>,
        path: impl Into<Vector<3>>,
        core: &mut Core,
    ) -> ShellExtendedBySweep;
}

Sweep a Face that is part of a Shell

See module documentation for more information.

Required Methods

fn sweep_face_of_shell( &self, face: Handle<Face>, path: impl Into<Vector<3>>, core: &mut Core, ) -> ShellExtendedBySweep

Sweep the Face of the Shell

Extends the shell, adding the new faces to it.

Limitation

When generating new faces, these must NOT coincide with any existing faces in the shell.

Panics

Panics, if the face has interior cycles. This is not a fundamental limitation, but none the less not yet supported.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.