fj_core::operations::transform

Trait TransformObject

pub trait TransformObject: Sized {
    type Transformed;

    // Required method
    fn transform_with_cache(
        self,
        transform: &Transform,
        core: &mut Core,
        cache: &mut TransformCache,
    ) -> Self::Transformed;

    // Provided methods
    fn transform(
        self,
        transform: &Transform,
        core: &mut Core,
    ) -> Self::Transformed { ... }
    fn translate(
        self,
        offset: impl Into<Vector<3>>,
        core: &mut Core,
    ) -> Self::Transformed { ... }
    fn rotate(
        self,
        axis_angle: impl Into<Vector<3>>,
        core: &mut Core,
    ) -> Self::Transformed { ... }
}

Transform an object

Implementation Note

So far, a general transform method is available, along some convenience methods for more specific transformations.

More convenience methods can be added as required. The only reason this hasn’t been done so far, is that no one has put in the work yet.

Required Associated Types

type Transformed

The result of the transformation

Required Methods

fn transform_with_cache( self, transform: &Transform, core: &mut Core, cache: &mut TransformCache, ) -> Self::Transformed

Transform the object using the provided cache

Provided Methods

fn transform(self, transform: &Transform, core: &mut Core) -> Self::Transformed

Transform the object

fn translate( self, offset: impl Into<Vector<3>>, core: &mut Core, ) -> Self::Transformed

Translate the object

Convenience wrapper around TransformObject::transform.

fn rotate( self, axis_angle: impl Into<Vector<3>>, core: &mut Core, ) -> Self::Transformed

Rotate the object

Convenience wrapper around TransformObject::transform.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.