fj_core::storage

Struct Handle

pub struct Handle<T> { /* private fields */ }

A handle that references a stored object

You can get an instance of Handle by inserting an object into a store. A handle dereferences to the object it points to, via its Deref implementation.

Bare objects and stored objects

A bare object is just that: an instance of a bare object type. Once a bare objects is inserted into storage, it becomes a stored object. A stored object is owned by the store, and can be referenced through instances of Handle.

The point of doing this, is to provide objects with a unique identity, via their location within storage. The importance of this is expanded upon in the next section.

Equality and Identity

Most objects have Eq/PartialEq implementations that can be used to determine equality. Those implementations are derived, meaning two objects are equal, if all of their fields are equal. This can be used to compare objects structurally. Handle’s own Eq/PartialEq implementations defer to those of the stored object it references.

However, that two objects are equal does not mean they are identical.

This distinction is relevant, because non-identical objects that are supposed to be equal can in fact end up equal, if they are created based on simple input data (as you might have in a unit test). But they might end up slightly different, if they are created based on complex input data (as you might have in a real-world scenario). This situation would most likely result in a bug that is not easily caught in testing.

You can compare the identity of two Handles, by comparing the values returned by Handle::id.

Validation Must Use Identity

To prevent situations where everything looks fine during development, but you end up with a bug in production, any validation code that compares objects and expects them to be the same, must do that comparison based on identity, not equality. That way, this problem can never happen, because we never expect non-identical objects to be equal.

Implementations

impl<T> Handle<T>

pub fn id(&self) -> ObjectId

Access the object’s unique id

pub fn clone_object(&self) -> T
where T: Clone,

Return a bare object, which is a clone of the referenced stored object