fj_core

Module topology

Objects of a shape

Objects, in Fornjot parlance, are the elements that make up shapes. Objects can reference each other, forming a directed acyclic graph (DAG).

There are two top-level objects (Sketch and Solid) which represent whole shapes (2D and 3D, respectively), while all other objects are referenced (directly or indirectly) by these top-level objects.

All objects are stored in centralized storage (see Topology) and referred to through a Handle.

Structs

AboutToBeStored

Implementation of Form for objects that are about to be stored

Bare

Implementation of Form for bare objects

Curve

A curve

Cycle

A cycle of connected edges

Face

A face of a shape

HalfEdge

A directed half-edge, defined in a surface’s 2D space

ObjectSet

An ordered set of objects

Region

A single, continuous 2d region; may contain holes

Shell

A 3-dimensional closed shell

Sketch

A 2-dimensional shape

Solid

A 3-dimensional shape, built from Shells. Many Solids will contains only one shell, but if the Solid contains cavities they will be represented by a shell each, as well as a shell for the outside.

Stored

Implementation of Form for stored objects

Surface

A two-dimensional shape

Surfaces

Store for Surfaces

Topology

The stores for all topological objects

Vertex

A vertex that identifies a point in space

Enums

AnyObject

An enum that can hold any object

Handedness

The handedness of a face’s coordinate system

Traits

Form

The form that an object can take

IsObject

A trait implemented for all object types

Type Aliases

ObjectSetIntoIter

An owned iterator over an ObjectSet

ObjectSetIter

An borrowed iterator over an ObjectSet