fj_core::topology

Enum AnyObject

pub enum AnyObject<F: Form> {
    Curve(F::Form<Curve>),
    Cycle(F::Form<Cycle>),
    Face(F::Form<Face>),
    HalfEdge(F::Form<HalfEdge>),
    Region(F::Form<Region>),
    Shell(F::Form<Shell>),
    Sketch(F::Form<Sketch>),
    Solid(F::Form<Solid>),
    Surface(F::Form<Surface>),
    Vertex(F::Form<Vertex>),
}

An enum that can hold any object

This enum is generic over the form that the object takes. An AnyObject<Bare> contains a bare objects, for example Curve. An AnyObject<Stored> contains a handle referencing a stored object, for example Handle<Curve>.

Variants

Curve(F::Form<Curve>)

A curve

Cycle(F::Form<Cycle>)

A cycle

Face(F::Form<Face>)

A face

HalfEdge(F::Form<HalfEdge>)

A half-edge

Region(F::Form<Region>)

A region

Shell(F::Form<Shell>)

A shell

Sketch(F::Form<Sketch>)

A sketch

Solid(F::Form<Solid>)

A solid

Surface(F::Form<Surface>)

A surface

Vertex(F::Form<Vertex>)

A vertex

Implementations

impl AnyObject<Stored>

pub fn id(&self) -> ObjectId

Access the ID of the object

pub fn validate( &self, config: &ValidationConfig, errors: &mut Vec<ValidationError>, geometry: &Geometry, )

Validate the object with a pre-defined validation configuration

impl AnyObject<AboutToBeStored>

pub fn insert(self, topology: &mut Topology) -> AnyObject<Stored>

Insert the object into its respective store