fj_core::topology

Struct Cycle

pub struct Cycle { /* private fields */ }

A cycle of connected edges

Implementations

impl Cycle

pub fn new(half_edges: impl IntoIterator<Item = Handle<HalfEdge>>) -> Self

Create an instance of Cycle

pub fn half_edges(&self) -> &ObjectSet<HalfEdge>

Access the edges that make up the cycle

pub fn winding(&self, geometry: &Geometry, surface: &Handle<Surface>) -> Winding

Indicate the cycle’s winding, assuming a right-handed coordinate system

Please note that this is not the winding of the cycle, only one of the two possible windings, depending on the direction you look at the surface that the cycle is defined on from.