fj_core::topology

Struct Region

pub struct Region { /* private fields */ }

A single, continuous 2d region; may contain holes

Interior cycles must have the opposite winding of the exterior cycle, meaning on the front side of the region, they must appear clockwise. This means that all HalfEdges that bound a Region have the interior of the region on their left side (on the region’s front side).

Implementations

impl Region

pub fn new( exterior: Handle<Cycle>, interiors: impl IntoIterator<Item = Handle<Cycle>>, ) -> Self

Construct an instance of Region

pub fn exterior(&self) -> &Handle<Cycle>

Access the cycle that bounds the region on the outside

pub fn interiors(&self) -> &ObjectSet<Cycle>

Access the cycles that bound the region on the inside

Each of these cycles defines a hole in the region .

pub fn all_cycles(&self) -> impl Iterator<Item = &Handle<Cycle>>

Access all cycles of the region (both exterior and interior)