fj_core::topology

Struct Solid

pub struct Solid { /* private fields */ }

A 3-dimensional shape, built from Shells. Many Solids will contains only one shell, but if the Solid contains cavities they will be represented by a shell each, as well as a shell for the outside.

Implementation Note

The shells that form the boundaries of the solid must not intersect. This is not currently validated.

Implementations

impl Solid

pub fn new(shells: impl IntoIterator<Item = Handle<Shell>>) -> Self

Construct an empty instance of Solid

pub fn shells(&self) -> &ObjectSet<Shell>

Access the solid’s shells