fj_core::topology

Trait IsObject

pub trait IsObject {
    type BareObject;
}

A trait implemented for all object types

This trait is implemented for both bare objects (T) and stored objects (Handle<T>). The BareObject associated type provides access to the bare object type.

This is a piece of infrastructure that is useful for other traits, which would otherwise have to duplicate its functionality. Users are unlikely to be affected by this trait.

Required Associated Types

type BareObject

The type of the bare object