fj_core::topology

Type Alias ObjectSetIntoIter

pub type ObjectSetIntoIter<T> = IntoIter<Handle<T>>;

An owned iterator over an ObjectSet

You might wonder why we’re returning references to Handles here, when Handle already is a kind of reference, and easily cloned.

Most of the time that doesn’t make a difference, but there are use cases where dealing with owned Handles is inconvenient, for example when using iterator adapters. You can’t return a reference to the argument of an adapter’s closure, if you own that argument. You can, if you just reference the argument.

Aliased Type

pub struct ObjectSetIntoIter<T> { /* private fields */ }