fj_core

Module validation

Infrastructure for validating objects

Nomenclature

Validation is the process of checking that objects meet specific requirements. Each kind of object has its own set of requirements.

An object that meets all the requirement for its kind is considered valid. An object that does not meet all of them is considered invalid. This results in a validation error, which is represented by ValidationError.

Implementation Note

This is a new module whose goal is to replace crate::validate. While this transition is ongoing, both modules will be incomplete.

Issues that track the transition:

https://github.com/hannobraun/fornjot/issues/1713

https://github.com/hannobraun/fornjot/issues/2157

Modules

checks

All validation checks

Structs

Validation

Errors that occurred while validating the objects inserted into the stores

ValidationConfig

Configuration required for the validation process

ValidationErrors

A collection of validation errors

Enums

ValidationError

An error that can occur during a validation

Traits

ValidationCheck

Run a specific validation check on an object